requirements-dev.txt
aio-pika==9.4.3
aiormq==6.8.1
astroid==3.3.5
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
dill==0.3.9
dnspython==2.7.0
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.25.0
idna==3.10
isort==5.13.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
py-real-esrgan==2.0.0
pylint==3.3.1
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
sympy==1.13.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.17.0
